feat: restructure project by organizing types and adding new input and execution classes
