+++ src/acodex/constants/toml.py
@@ -1,0 +1,3 @@
+import re
+
+TOML_BARE_KEY = re.compile(r"^[A-Za-z0-9_-]+$")
