+++ src/acodex/types/thread_options.py
